requirements.txt
flask
flask-login
flask-sqlalchemy
psycopg2
